+++ src/components/game/live_game.rs
@@ -1,12 +1,11 @@
 use crate::components::boxscore::TeamBatterBoxscore;
 use crate::components::constants::TEAM_IDS;
-use crate::components::game::at_bat::AtBatV2;
-use crate::components::game::matchup::Player;
+use crate::components::game::at_bat::AtBat;
+use crate::components::game::player::{Player, PlayerStats};
 use crate::components::game::win_probability::WinProbability;
 use crate::components::linescore::LineScore;
 use crate::components::standings::Team;
 use indexmap::IndexMap;
-use mlb_api::boxscore::Player as ApiPlayer;
 use mlb_api::live::LiveResponse;
 use mlb_api::plays::Play;
 use mlb_api::win_probability::WinProbabilityResponse;
@@ -17,57 +16,24 @@ pub type AtBatIndex = u8;
 pub type PlayerId = u64;
 pub type PlayerMap = HashMap<PlayerId, Player>;
 
-static DEFAULT_AT_BAT: LazyLock<AtBatV2> = LazyLock::new(AtBatV2::default);
+static DEFAULT_AT_BAT: LazyLock<AtBat> = LazyLock::new(AtBat::default);
 
 #[derive(Default)]
-pub struct GameStateV2 {
+pub struct GameState {
     pub game_id: u64,
     pub home_team: Team,
     pub away_team: Team,
     pub linescore: LineScore,
     pub boxscore: TeamBatterBoxscore,
     pub current_at_bat: AtBatIndex,
-    pub at_bats: IndexMap<AtBatIndex, AtBatV2>,
+    pub at_bats: IndexMap<AtBatIndex, AtBat>,
     pub on_deck: Option<PlayerId>,
     pub in_hole: Option<PlayerId>,
     pub win_probability: WinProbability,
     pub players: PlayerMap,
 }
 
-#[derive(Default, Debug)]
-pub struct PlayerStats {
-    pub summary: Option<String>,
-    pub note: Option<String>,
-    pub pitches_thrown: Option<u8>,
-    pub strikes: Option<u8>,
-    #[allow(dead_code)]
-    pub balls: Option<u8>,
-}
-
-impl From<&ApiPlayer> for PlayerStats {
-    fn from(person: &ApiPlayer) -> Self {
-        let is_pitcher = person.position.position_type == "Pitcher";
-        if is_pitcher {
-            Self {
-                summary: person.stats.pitching.summary.clone(),
-                note: person.stats.pitching.note.clone(),
-                pitches_thrown: person.stats.pitching.pitches_thrown.map(|p| p as u8),
-                strikes: person.stats.pitching.strikes.map(|p| p as u8),
-                balls: person.stats.pitching.balls.map(|p| p as u8),
-            }
-        } else {
-            Self {
-                summary: person.stats.batting.summary.clone(),
-                note: person.stats.batting.note.clone(),
-                pitches_thrown: None,
-                strikes: None,
-                balls: None,
-            }
-        }
-    }
-}
-
-impl GameStateV2 {
+impl GameState {
     /// Update with latest data from the API.
     pub fn update(&mut self, live_data: &LiveResponse, win_probability: &WinProbabilityResponse) {
         if self.game_id != live_data.game_pk {
@@ -138,20 +104,20 @@ impl GameStateV2 {
     }
 
     /// Will always return an at bat. If there isn't one, it will return the default.
-    pub fn get_latest_at_bat(&self) -> &AtBatV2 {
+    pub fn get_latest_at_bat(&self) -> &AtBat {
         self.at_bats
             .get(&self.current_at_bat)
             .unwrap_or_else(|| &DEFAULT_AT_BAT)
     }
 
     /// May not exist.
-    pub fn get_at_bat_by_index(&self, index: u8) -> Option<&AtBatV2> {
+    pub fn get_at_bat_by_index(&self, index: u8) -> Option<&AtBat> {
         self.at_bats.get(&index)
     }
 
     /// Helper function to try to get an at bat by index. If it doesn't exist, it will return the
     /// latest at bat. It will also return `true` if the at bat is the current at bat.
-    pub fn get_at_bat_by_index_or_current(&self, index: Option<u8>) -> (&AtBatV2, bool) {
+    pub fn get_at_bat_by_index_or_current(&self, index: Option<u8>) -> (&AtBat, bool) {
         let idx = index.unwrap_or(self.current_at_bat);
         match self.get_at_bat_by_index(idx) {
             Some(at_bat) => (at_bat, idx == self.current_at_bat),
@@ -175,7 +141,7 @@ impl GameStateV2 {
 
     /// Useful for updating current play.
     pub fn update_single_play(&mut self, play: &Play) {
-        let at_bat = AtBatV2::from(play);
+        let at_bat = AtBat::from(play);
         self.at_bats.insert(at_bat.index, at_bat);
     }
